+++ src/common/hist_util.h
@@ -454,182 +454,6 @@ class HistCollection {
   std::vector<size_t> row_ptr_;
 };
 
-/*!
- * \brief Stores temporary histograms to compute them in parallel
- * Supports processing multiple tree-nodes for nested parallelism
- * Able to reduce histograms across threads in efficient way
- */
-class ParallelGHistBuilder {
- public:
-  void Init(size_t nbins) {
-    if (nbins != nbins_) {
-      hist_buffer_.Init(nbins);
-      nbins_ = nbins;
-    }
-  }
-
-  // Add new elements if needed, mark all hists as unused
-  // targeted_hists - already allocated hists which should contain final results after Reduce() call
-  void Reset(size_t nthreads, size_t nodes, const BlockedSpace2d& space,
-             const std::vector<GHistRow>& targeted_hists) {
-    hist_buffer_.Init(nbins_);
-    tid_nid_to_hist_.clear();
-    threads_to_nids_map_.clear();
-
-    targeted_hists_ = targeted_hists;
-
-    CHECK_EQ(nodes, targeted_hists.size());
-
-    nodes_    = nodes;
-    nthreads_ = nthreads;
-
-    MatchThreadsToNodes(space);
-    AllocateAdditionalHistograms();
-    MatchNodeNidPairToHist();
-
-    hist_was_used_.resize(nthreads * nodes_);
-    std::fill(hist_was_used_.begin(), hist_was_used_.end(), static_cast<int>(false));
-  }
-
-  // Get specified hist, initialize hist by zeros if it wasn't used before
-  GHistRow GetInitializedHist(size_t tid, size_t nid) {
-    CHECK_LT(nid, nodes_);
-    CHECK_LT(tid, nthreads_);
-
-    int idx = tid_nid_to_hist_.at({tid, nid});
-    if (idx >= 0) {
-      hist_buffer_.AllocateData(idx);
-    }
-    GHistRow hist = idx == -1 ? targeted_hists_[nid] : hist_buffer_[idx];
-
-    if (!hist_was_used_[tid * nodes_ + nid]) {
-      InitilizeHistByZeroes(hist, 0, hist.size());
-      hist_was_used_[tid * nodes_ + nid] = static_cast<int>(true);
-    }
-
-    return hist;
-  }
-
-  // Reduce following bins (begin, end] for nid-node in dst across threads
-  void ReduceHist(size_t nid, size_t begin, size_t end) const {
-    CHECK_GT(end, begin);
-    CHECK_LT(nid, nodes_);
-
-    GHistRow dst = targeted_hists_[nid];
-
-    bool is_updated = false;
-    for (size_t tid = 0; tid < nthreads_; ++tid) {
-      if (hist_was_used_[tid * nodes_ + nid]) {
-        is_updated = true;
-
-        int idx = tid_nid_to_hist_.at({tid, nid});
-        GHistRow src = idx == -1 ? targeted_hists_[nid] : hist_buffer_[idx];
-
-        if (dst.data() != src.data()) {
-          IncrementHist(dst, src, begin, end);
-        }
-      }
-    }
-    if (!is_updated) {
-      // In distributed mode - some tree nodes can be empty on local machines,
-      // So we need just set local hist by zeros in this case
-      InitilizeHistByZeroes(dst, begin, end);
-    }
-  }
-
-  void MatchThreadsToNodes(const BlockedSpace2d& space) {
-    const size_t space_size = space.Size();
-    const size_t chunck_size = space_size / nthreads_ + !!(space_size % nthreads_);
-
-    threads_to_nids_map_.resize(nthreads_ * nodes_, false);
-
-    for (size_t tid = 0; tid < nthreads_; ++tid) {
-      size_t begin = chunck_size * tid;
-      size_t end   = std::min(begin + chunck_size, space_size);
-
-      if (begin < space_size) {
-        size_t nid_begin = space.GetFirstDimension(begin);
-        size_t nid_end   = space.GetFirstDimension(end-1);
-
-        for (size_t nid = nid_begin; nid <= nid_end; ++nid) {
-          // true - means thread 'tid' will work to compute partial hist for node 'nid'
-          threads_to_nids_map_[tid * nodes_ + nid] = true;
-        }
-      }
-    }
-  }
-
-  void AllocateAdditionalHistograms() {
-    size_t hist_allocated_additionally = 0;
-
-    for (size_t nid = 0; nid < nodes_; ++nid) {
-      int nthreads_for_nid = 0;
-
-      for (size_t tid = 0; tid < nthreads_; ++tid) {
-        if (threads_to_nids_map_[tid * nodes_ + nid]) {
-          nthreads_for_nid++;
-        }
-      }
-
-      // In distributed mode - some tree nodes can be empty on local machines,
-      // set nthreads_for_nid to 0 in this case.
-      // In another case - allocate additional (nthreads_for_nid - 1) histograms,
-      // because one is already allocated externally (will store final result for the node).
-      hist_allocated_additionally += std::max<int>(0, nthreads_for_nid - 1);
-    }
-
-    for (size_t i = 0; i < hist_allocated_additionally; ++i) {
-      hist_buffer_.AddHistRow(i);
-    }
-  }
-
-  [[nodiscard]] bst_bin_t TotalBins() const { return nbins_; }
-
- private:
-  void MatchNodeNidPairToHist() {
-    size_t hist_allocated_additionally = 0;
-
-    for (size_t nid = 0; nid < nodes_; ++nid) {
-      bool first_hist = true;
-      for (size_t tid = 0; tid < nthreads_; ++tid) {
-        if (threads_to_nids_map_[tid * nodes_ + nid]) {
-          if (first_hist) {
-            tid_nid_to_hist_[{tid, nid}] = -1;
-            first_hist = false;
-          } else {
-            tid_nid_to_hist_[{tid, nid}] = hist_allocated_additionally++;
-          }
-        }
-      }
-    }
-  }
-
-  /*! \brief number of bins in each histogram */
-  size_t nbins_ = 0;
-  /*! \brief number of threads for parallel computation */
-  size_t nthreads_ = 0;
-  /*! \brief number of nodes which will be processed in parallel  */
-  size_t nodes_ = 0;
-  /*! \brief Buffer for additional histograms for Parallel processing  */
-  HistCollection hist_buffer_;
-  /*!
-   * \brief Marks which hists were used, it means that they should be merged.
-   * Contains only {true or false} values
-   * but 'int' is used instead of 'bool', because std::vector<bool> isn't thread safe
-   */
-  std::vector<int> hist_was_used_;
-
-  /*! \brief Buffer for additional histograms for Parallel processing  */
-  std::vector<bool> threads_to_nids_map_;
-  /*! \brief Contains histograms for final results  */
-  std::vector<GHistRow> targeted_hists_;
-  /*!
-   * \brief map pair {tid, nid} to index of allocated histogram from hist_buffer_ and targeted_hists_,
-   * -1 is reserved for targeted_hists_
-   */
-  std::map<std::pair<size_t, size_t>, int> tid_nid_to_hist_;
-};
-
 // construct a histogram via histogram aggregation
 template <bool any_missing>
 void BuildHist(Span<GradientPair const> gpair, const RowSetCollection::Elem row_indices,
